# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  biome?: string | null;
  phase: string;
  /** @nullable */
  solarScore?: number | null;
  /** @nullable */
  embodiedCarbon?: number | null;
  /** @nullable */
  waterHarvesting?: number | null;
  /** @nullable */
  estimatedCost?: number | null;
  status: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  biome?: string;
  phase?: string;
  estimatedCost?: number;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  biome?: string;
  phase?: string;
  solarScore?: number;
  embodiedCarbon?: number;
  waterHarvesting?: number;
  estimatedCost?: number;
  status?: string;
  thumbnailUrl?: string;
}

export interface ProjectStats {
  totalProjects: number;
  activeProjects: number;
  totalCarbonSaved: number;
  avgSolarScore: number;
  recentProjects: Project[];
}

export interface Material {
  id: number;
  name: string;
  category: string;
  description: string;
  embodiedCarbon: number;
  localAvailability: string;
  durabilityYears: number;
  recyclability: number;
  isFeatured: boolean;
  tags: string[];
  /** @nullable */
  imageUrl?: string | null;
}

export interface MoodboardInput {
  /** Text description of desired vibe, biome, or inspiration */
  description: string;
  /** Optional URLs of uploaded inspiration images */
  imageUrls?: string[];
}

export type SolaraSpecParametricHints = {
  /** @nullable */
  floorAreaSqm: number | null;
  /** @nullable */
  stories: number | null;
  /** @nullable */
  roofType: string | null;
  /** @nullable */
  primaryStructure: string | null;
};

export interface SolaraSpec {
  title: string;
  biome: string;
  /** Hex color palette derived from the moodboard */
  palette: string[];
  tags: string[];
  summary: string;
  suggestedMaterials: string[];
  parametricHints: SolaraSpecParametricHints;
  estimatedCarbon: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListMaterialsParams = {
category?: string;
maxCarbon?: number;
};