# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SolaraForge API — AI-powered solarpunk regenerative habitat designer
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Habitat design projects
  - name: materials
    description: Regenerative materials library
  - name: moodboard
    description: Moodboard studio and AI concept analysis
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Projects ────────────────────────────────────────────
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new habitat project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Dashboard stats — project counts, carbon totals, recent activity
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ── Materials ────────────────────────────────────────────
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List regenerative materials library
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: maxCarbon
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"

  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /materials/featured:
    get:
      operationId: getFeaturedMaterials
      tags: [materials]
      summary: Get featured/top-rated regenerative materials
      responses:
        "200":
          description: Featured materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"

  # ── Moodboard ─────────────────────────────────────────────
  /moodboard/analyze:
    post:
      operationId: analyzeMoodboard
      tags: [moodboard]
      summary: Analyze a moodboard description and return a SolaraSpec concept card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodboardInput"
      responses:
        "200":
          description: Generated SolaraSpec concept card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolaraSpec"

  # ── OpenAI Chat ───────────────────────────────────────────
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Project schemas ─────────────────────────────────────
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        biome:
          type: ["string", "null"]
        phase:
          type: string
        solarScore:
          type: ["number", "null"]
        embodiedCarbon:
          type: ["number", "null"]
        waterHarvesting:
          type: ["number", "null"]
        estimatedCost:
          type: ["number", "null"]
        status:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phase
        - status
        - createdAt
        - updatedAt

    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        biome:
          type: string
        phase:
          type: string
        estimatedCost:
          type: number
      required:
        - name

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        biome:
          type: string
        phase:
          type: string
        solarScore:
          type: number
        embodiedCarbon:
          type: number
        waterHarvesting:
          type: number
        estimatedCost:
          type: number
        status:
          type: string
        thumbnailUrl:
          type: string

    ProjectStats:
      type: object
      properties:
        totalProjects:
          type: integer
        activeProjects:
          type: integer
        totalCarbonSaved:
          type: number
        avgSolarScore:
          type: number
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required:
        - totalProjects
        - activeProjects
        - totalCarbonSaved
        - avgSolarScore
        - recentProjects

    # ── Material schemas ────────────────────────────────────
    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        embodiedCarbon:
          type: number
        localAvailability:
          type: string
        durabilityYears:
          type: integer
        recyclability:
          type: number
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - category
        - description
        - embodiedCarbon
        - localAvailability
        - durabilityYears
        - recyclability
        - isFeatured
        - tags

    # ── Moodboard schemas ───────────────────────────────────
    MoodboardInput:
      type: object
      properties:
        description:
          type: string
          description: Text description of desired vibe, biome, or inspiration
        imageUrls:
          type: array
          items:
            type: string
          description: Optional URLs of uploaded inspiration images
      required:
        - description

    SolaraSpec:
      type: object
      properties:
        title:
          type: string
        biome:
          type: string
        palette:
          type: array
          items:
            type: string
          description: Hex color palette derived from the moodboard
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
        suggestedMaterials:
          type: array
          items:
            type: string
        parametricHints:
          type: object
          properties:
            floorAreaSqm:
              type: ["number", "null"]
            stories:
              type: ["integer", "null"]
            roofType:
              type: ["string", "null"]
            primaryStructure:
              type: ["string", "null"]
          required:
            - floorAreaSqm
            - stories
            - roofType
            - primaryStructure
        estimatedCarbon:
          type: number
      required:
        - title
        - biome
        - palette
        - tags
        - summary
        - suggestedMaterials
        - parametricHints
        - estimatedCarbon

    # ── OpenAI schemas ──────────────────────────────────────
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error