# lib/api-zod/src/generated/types/moodboardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */

export interface MoodboardInput {
  /** Text description of desired vibe, biome, or inspiration */
  description: string;
  /** Optional URLs of uploaded inspiration images */
  imageUrls?: string[];
}