# lib/api-zod/src/generated/types/openaiConversationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversationInput {
  title: string;
}