# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';

export interface ProjectStats {
  totalProjects: number;
  activeProjects: number;
  totalCarbonSaved: number;
  avgSolarScore: number;
  recentProjects: Project[];
}