# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  name?: string;
  description?: string;
  biome?: string;
  phase?: string;
  solarScore?: number;
  embodiedCarbon?: number;
  waterHarvesting?: number;
  estimatedCost?: number;
  status?: string;
  thumbnailUrl?: string;
}