# lib/api-zod/src/generated/types/solaraSpec.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SolaraForge API — AI-powered solarpunk regenerative habitat designer
 * OpenAPI spec version: 0.1.0
 */
import type { SolaraSpecParametricHints } from './solaraSpecParametricHints';

export interface SolaraSpec {
  title: string;
  biome: string;
  /** Hex color palette derived from the moodboard */
  palette: string[];
  tags: string[];
  summary: string;
  suggestedMaterials: string[];
  parametricHints: SolaraSpecParametricHints;
  estimatedCarbon: number;
}